requirements.txt
astroid==2.3.3
click==7.1.1
colorama==0.4.3
Flask==1.1.2
Flask-WTF==0.14.3
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
python-dateutil==2.8.1
six==1.14.0
Werkzeug==1.0.1
wrapt==1.11.2
WTForms==2.2.1

